+++ NEWS.md
@@ -8,7 +8,7 @@
   no duplicate time points per subject, and this is explained also in the
   function documentation and the introduction vignette.
 - Previously in `mmrm` calls, the `weights` object in the environment where formula is
-  defined will be replaced by the weight used. Now this behaviour is removed and your variable
+  defined will be replaced by the weight used. Now this behavior is removed and your variable
   `weights` will not be replaced.
 ### Miscellaneous