+++ NEWS.md
@@ -9,7 +9,7 @@
   function documentation and the introduction vignette.
 - Previously in `mmrm` calls, the `weights` object in the environment where formula is
   defined will be replaced by the `weights` used internally. Now this behavior is removed and your variable
-  `weights` will not be replaced.
+  `weights` e.g. in the global environment will not be replaced.
 ### Miscellaneous
 
 - Deprecated `free_cores()` in favor of `parallelly::availableCores(omit = 1)`.